+++ server/src/main/java/org/allaymc/server/container/processor/DropActionProcessor.java
@@ -26,8 +26,8 @@ public class DropActionProcessor implements ContainerActionProcessor<DropAction>
         var slot = ContainerActionProcessor.fromNetworkSlotIndex(container, action.getSource().getSlot());
 
         var item = container.getItemStack(slot);
-        if (failToValidateStackNetworkId(item.getStackNetworkId(), action.getSource().getStackNetworkId())) {
-            log.warn("mismatch stack network id!");
+        if (failToValidateStackUniqueId(item.getUniqueId(), action.getSource().getStackNetworkId())) {
+            log.warn("mismatch stack unique id!");
             return error();
         }
 
@@ -53,7 +53,7 @@ public class DropActionProcessor implements ContainerActionProcessor<DropAction>
                                                 ContainerActionProcessor.toNetworkSlotIndex(container, slot),
                                                 ContainerActionProcessor.toNetworkSlotIndex(container, slot),
                                                 item.getCount(),
-                                                item.getStackNetworkId(),
+                                                item.getUniqueId(),
                                                 item.getCustomName(),
                                                 item.getDamage(),
                                                 ""